feat(fhirpath): iif criterion type validation (plan 2.3 wave 18)
Per FHIRPath spec §8.7.3, iif() criterion must be Boolean or empty;
non-boolean scalars raise TypeError and multi-item collections raise
EvaluationError. Fixes testCollectionBoolean1 and testIif6.

+++ crates/rh-fhirpath/src/evaluator/operations/comparison.rs
@@ -365,7 +365,7 @@ fn quantity_equivalent(a: f64, b: f64) -> bool {
     if prec == 0 {
         return (a - b).abs() < 1.0;
     }
-    let scale = 10f64.powi(prec as i32 - 1 - a.abs().log10().floor() as i32);
+    let scale = 10f64.powi(prec - 1 - a.abs().log10().floor() as i32);
     (a * scale).round() == (b * scale).round()
 }
